feat(select)!: compile with tsup
BREAKING CHANGE: import resources from the root file instead of @availity/select/resources
