--- packages/select/index.d.ts
@@ -1,3 +1,0 @@
-export { default, SelectProps } from './types/Select';
-export { default as SelectField, SelectFieldProps } from './types/SelectField';
-export { default as ResourceSelect, ResourceSelectProps } from './types/ResourceSelect';
