--- packages/select/index.js
@@ -1,3 +1,0 @@
-export { default } from './src/Select';
-export { default as SelectField } from './src/SelectField';
-export { default as ResourceSelect } from './src/ResourceSelect';
