+++ packages/select/jest.config.js
@@ -1,0 +1,7 @@
+const global = require('../../jest/global-config');
+
+module.exports = {
+  ...global,
+  displayName: 'select',
+  coverageDirectory: '../../coverage/select',
+};
