+++ packages/select/package.json
@@ -18,25 +18,36 @@
   },
   "license": "MIT",
   "author": "Evan Sharp <evan.sharp@availity.com>",
-  "main": "index.js",
-  "types": "index.d.ts",
+  "browser": "./dist/index.js",
+  "main": "./dist/index.js",
+  "module": "./dist/index.mjs",
+  "types": "./dist/index.d.ts",
+  "scripts": {
+    "build": "tsup src/index.js --format esm,cjs --loader \".js=jsx\" --dts",
+    "dev": "tsup src/index.js --format esm,cjs --loader \".js=jsx\" --watch --dts",
+    "clean": "rm -rf node_modules && rm -rf dist",
+    "bundlesize": "bundlesize",
+    "publish": "yarn npm publish --tolerate-republish --access public"
+  },
   "dependencies": {
     "classnames": "^2.3.1",
     "lodash": "^4.17.21",
-    "prop-types": "^15.7.2",
-    "qs": "^6.5.2",
-    "react-select": "^5.0.0",
-    "react-select-async-paginate": "^0.6.0"
+    "prop-types": "^15.8.1",
+    "qs": "^6.10.3",
+    "react-select": "^5.3.2",
+    "react-select-async-paginate": "^0.6.1"
   },
   "devDependencies": {
-    "@availity/api-axios": "^6.0.0",
-    "@availity/form": "^1.3.1",
-    "axios": "^0.21.1",
-    "formik": "^2.2.8",
+    "@availity/api-axios": "^7.0.1",
+    "@availity/form": "workspace:*",
+    "axios": "^0.21.4",
+    "formik": "^2.2.9",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "reactstrap": "^8.9.0",
-    "yup": "^0.30.0"
+    "reactstrap": "^8.10.1",
+    "tsup": "^5.12.7",
+    "typescript": "^4.6.4",
+    "yup": "^0.32.11"
   },
   "peerDependencies": {
     "@availity/api-axios": "^6.0.0",
@@ -44,7 +55,7 @@
     "axios": "^0.21.1",
     "formik": "^2.0.1",
     "react": ">=16.8.3",
-    "reactstrap": "^8.0.0 || ^7.0.0 || ^6.0.0"
+    "reactstrap": "^8.0.0"
   },
   "publishConfig": {
     "access": "public"
