+++ packages/select/src/index.d.ts
@@ -1,0 +1,4 @@
+export { default, SelectProps } from './Select';
+export { default as SelectField, SelectFieldProps } from './SelectField';
+export { default as ResourceSelect, ResourceSelectProps } from './ResourceSelect';
+export * from './resources';
