+++ packages/select/src/index.js
@@ -1,0 +1,4 @@
+export { default } from './Select';
+export { default as SelectField } from './SelectField';
+export { default as ResourceSelect } from './ResourceSelect';
+export * from './resources';
